docs(a11y): document keyboard shortcuts and focus model
