+++ src/app/milestones/__tests__/page.test.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { render, screen, waitFor, act, within } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-import MilestonesPage, { SAMPLE_MILESTONES, SAMPLE_DISMISSED_KEY } from '../page';
+import MilestonesPage from '../page';
+import { SAMPLE_MILESTONES, SAMPLE_DISMISSED_KEY } from '../constants';
 import { listMilestones, saveMilestone } from '@/lib/repository';
 import type { Milestone } from '@/types/domain';
