+++ src/app/milestones/constants.ts
@@ -1,0 +1,46 @@
+import type { Milestone } from '@/types/domain';
+
+export const SAMPLE_DISMISSED_KEY = 'talenttrust-milestones-sample-dismissed';
+
+export const SAMPLE_MILESTONES: Milestone[] = [
+  {
+    id: '1',
+    title: 'Project Kickoff & Discovery',
+    status: 'Completed',
+    payout: 2500,
+    currency: 'USD',
+    dueDate: '2026-03-15',
+  },
+  {
+    id: '2',
+    title: 'UI/UX Design Handoff',
+    status: 'Paid',
+    payout: 3500,
+    currency: 'USD',
+    dueDate: '2026-04-01',
+  },
+  {
+    id: '3',
+    title: 'Frontend Development – Sprint 1',
+    status: 'Pending',
+    payout: 5000,
+    currency: 'USD',
+    dueDate: '2026-05-01',
+  },
+  {
+    id: '4',
+    title: 'API Integration & Testing',
+    status: 'Pending',
+    payout: 4000,
+    currency: 'USD',
+    dueDate: '2026-05-15',
+  },
+  {
+    id: '5',
+    title: 'Payment Gateway Integration',
+    status: 'Disputed',
+    payout: 3000,
+    currency: 'USD',
+    dueDate: '2026-04-20',
+  },
+];
